config tests for Linux
